chore: move rabbitmq tests to correct folder
