+++ BuildingBlocks/test/BuildingBlocks.Infrastructure.Tests/EventBus/RabbitMq/EventBusRabbitMqTests.cs
@@ -1,7 +1,7 @@
 ﻿using Backbone.BuildingBlocks.Infrastructure.EventBus.RabbitMQ;
 using Backbone.Modules.TestModule.Application.DomainEvents;
 
-namespace Backbone.BuildingBlocks.Infrastructure.Tests.EventBus.EventBusRabbitMQ;
+namespace Backbone.BuildingBlocks.Infrastructure.Tests.EventBus.RabbitMq;
 
 public class EventBusRabbitMqTests : AbstractTestsBase
 {
