+++ BuildingBlocks/test/BuildingBlocks.Infrastructure.Tests/EventBus/RabbitMq/EventExtensionsTests.cs
@@ -3,7 +3,7 @@ using Backbone.BuildingBlocks.Domain.Events;
 using Backbone.BuildingBlocks.Infrastructure.EventBus;
 using Backbone.Modules.TestModule.Application.DomainEvents;
 
-namespace Backbone.BuildingBlocks.Infrastructure.Tests.EventBus.EventBusRabbitMQ
+namespace Backbone.BuildingBlocks.Infrastructure.Tests.EventBus.RabbitMq
 {
     public class DomainEventNamingExtensionsTests : AbstractTestsBase
     {
